+++ src/bridge/mod.rs
@@ -1,6 +1,3 @@
-#[macro_use]
-pub mod layouts;
-
 mod events;
 mod handler;
 mod ui_commands;
@@ -23,7 +20,6 @@ use crate::settings::*;
 use crate::window::window_geometry_or_default;
 pub use events::*;
 use handler::NeovimHandler;
-pub use layouts::*;
 pub use ui_commands::UiCommand;
 
 #[cfg(windows)]
