+++ src/window/winit/layouts/qwerty.rs
@@ -1,0 +1,164 @@
+use super::unsupported_key;
+
+use skulpin::winit::event::VirtualKeyCode as Keycode;
+
+pub fn handle_qwerty_layout(
+    keycode: Keycode,
+    shift: bool,
+    ctrl: bool,
+    alt: bool,
+) -> Option<(&'static str, bool, bool, bool, bool)> {
+    match (keycode, shift, ctrl, alt) {
+        (Keycode::Back, shift, ctrl, alt) => Some(("BS", true, shift, ctrl, alt)),
+        (Keycode::Tab, shift, ctrl, alt) => Some(("Tab", true, shift, ctrl, alt)),
+        (Keycode::Return, shift, ctrl, alt) => Some(("Enter", true, shift, ctrl, alt)),
+        (Keycode::Escape, shift, ctrl, alt) => Some(("Esc", true, shift, ctrl, alt)),
+        (Keycode::Space, shift, ctrl, alt) => Some((" ", false, shift, ctrl, alt)),
+        (Keycode::Apostrophe, false, ctrl, alt) => Some(("'", false, false, ctrl, alt)),
+        (Keycode::Apostrophe, true, ctrl, alt) => Some(("\"", false, false, ctrl, alt)),
+        (Keycode::Comma, false, ctrl, alt) => Some((",", false, shift, ctrl, alt)),
+        (Keycode::Comma, true, ctrl, alt) => Some(("<", false, shift, ctrl, alt)),
+        (Keycode::Minus, false, ctrl, alt) => Some(("-", false, false, ctrl, alt)),
+        (Keycode::Minus, true, ctrl, alt) => Some(("_", false, false, ctrl, alt)),
+        (Keycode::Period, false, ctrl, alt) => Some((".", false, false, ctrl, alt)),
+        (Keycode::Period, true, ctrl, alt) => Some((">", false, false, ctrl, alt)),
+        (Keycode::Slash, false, ctrl, alt) => Some(("/", false, false, ctrl, alt)),
+        (Keycode::Slash, true, ctrl, alt) => Some(("?", false, false, ctrl, alt)),
+        (Keycode::Key0, false, ctrl, alt) => Some(("0", false, shift, ctrl, alt)),
+        (Keycode::Key0, true, ctrl, alt) => Some((")", true, shift, ctrl, alt)),
+        (Keycode::Key1, false, ctrl, alt) => Some(("1", false, shift, ctrl, alt)),
+        (Keycode::Key1, true, ctrl, alt) => Some(("!", true, shift, ctrl, alt)),
+        (Keycode::Key2, false, ctrl, alt) => Some(("2", false, false, ctrl, alt)),
+        (Keycode::Key2, true, ctrl, alt) => Some(("@", false, false, ctrl, alt)),
+        (Keycode::Key3, false, ctrl, alt) => Some(("3", false, false, ctrl, alt)),
+        (Keycode::Key3, true, ctrl, alt) => Some(("#", false, false, ctrl, alt)),
+        (Keycode::Key4, false, ctrl, alt) => Some(("4", false, false, ctrl, alt)),
+        (Keycode::Key4, true, ctrl, alt) => Some(("$", false, false, ctrl, alt)),
+        (Keycode::Key5, false, ctrl, alt) => Some(("5", false, false, ctrl, alt)),
+        (Keycode::Key5, true, ctrl, alt) => Some(("%", false, false, ctrl, alt)),
+        (Keycode::Key6, false, ctrl, alt) => Some(("6", false, false, ctrl, alt)),
+        (Keycode::Key6, true, ctrl, alt) => Some(("^", false, false, ctrl, alt)),
+        (Keycode::Key7, false, ctrl, alt) => Some(("7", false, false, ctrl, alt)),
+        (Keycode::Key7, true, ctrl, alt) => Some(("&", false, false, ctrl, alt)),
+        (Keycode::Key8, false, ctrl, alt) => Some(("8", false, false, ctrl, alt)),
+        (Keycode::Key8, true, ctrl, alt) => Some(("*", false, false, ctrl, alt)),
+        (Keycode::Key9, false, ctrl, alt) => Some(("9", false, false, ctrl, alt)),
+        (Keycode::Key9, true, ctrl, alt) => Some(("(", true, false, ctrl, alt)),
+        (Keycode::Colon, shift, ctrl, alt) => Some((":", false, shift, ctrl, alt)),
+        (Keycode::Semicolon, false, ctrl, alt) => Some((";", false, false, ctrl, alt)),
+        (Keycode::Semicolon, true, ctrl, alt) => Some((":", false, false, ctrl, alt)),
+        (Keycode::Equals, false, ctrl, alt) => Some(("=", false, false, ctrl, alt)),
+        (Keycode::Equals, true, ctrl, alt) => Some(("+", false, false, ctrl, alt)),
+        (Keycode::At, shift, ctrl, alt) => Some(("@", false, shift, ctrl, alt)),
+        (Keycode::LBracket, false, ctrl, alt) => Some(("[", false, false, ctrl, alt)),
+        (Keycode::LBracket, true, ctrl, alt) => Some(("{", false, false, ctrl, alt)),
+        (Keycode::Backslash, false, ctrl, alt) => Some(("\\", false, false, ctrl, alt)),
+        (Keycode::Backslash, true, ctrl, alt) => Some(("|", false, false, ctrl, alt)),
+        (Keycode::RBracket, false, ctrl, alt) => Some(("]", false, false, ctrl, alt)),
+        (Keycode::RBracket, true, ctrl, alt) => Some(("}", false, false, ctrl, alt)),
+        (Keycode::Caret, shift, ctrl, alt) => Some(("^", false, shift, ctrl, alt)),
+        (Keycode::Grave, false, ctrl, alt) => Some(("`", false, false, ctrl, alt)),
+        (Keycode::Grave, true, ctrl, alt) => Some(("~", false, false, ctrl, alt)),
+        (Keycode::A, shift, ctrl, alt) => Some(("a", false, shift, ctrl, alt)),
+        (Keycode::B, shift, ctrl, alt) => Some(("b", false, shift, ctrl, alt)),
+        (Keycode::C, shift, ctrl, alt) => Some(("c", false, shift, ctrl, alt)),
+        (Keycode::D, shift, ctrl, alt) => Some(("d", false, shift, ctrl, alt)),
+        (Keycode::E, shift, ctrl, alt) => Some(("e", false, shift, ctrl, alt)),
+        (Keycode::F, shift, ctrl, alt) => Some(("f", false, shift, ctrl, alt)),
+        (Keycode::G, shift, ctrl, alt) => Some(("g", false, shift, ctrl, alt)),
+        (Keycode::H, shift, ctrl, alt) => Some(("h", false, shift, ctrl, alt)),
+        (Keycode::I, shift, ctrl, alt) => Some(("i", false, shift, ctrl, alt)),
+        (Keycode::J, shift, ctrl, alt) => Some(("j", false, shift, ctrl, alt)),
+        (Keycode::K, shift, ctrl, alt) => Some(("k", false, shift, ctrl, alt)),
+        (Keycode::L, shift, ctrl, alt) => Some(("l", false, shift, ctrl, alt)),
+        (Keycode::M, shift, ctrl, alt) => Some(("m", false, shift, ctrl, alt)),
+        (Keycode::N, shift, ctrl, alt) => Some(("n", false, shift, ctrl, alt)),
+        (Keycode::O, shift, ctrl, alt) => Some(("o", false, shift, ctrl, alt)),
+        (Keycode::P, shift, ctrl, alt) => Some(("p", false, shift, ctrl, alt)),
+        (Keycode::Q, shift, ctrl, alt) => Some(("q", false, shift, ctrl, alt)),
+        (Keycode::R, shift, ctrl, alt) => Some(("r", false, shift, ctrl, alt)),
+        (Keycode::S, shift, ctrl, alt) => Some(("s", false, shift, ctrl, alt)),
+        (Keycode::T, shift, ctrl, alt) => Some(("t", false, shift, ctrl, alt)),
+        (Keycode::U, shift, ctrl, alt) => Some(("u", false, shift, ctrl, alt)),
+        (Keycode::V, shift, ctrl, alt) => Some(("v", false, shift, ctrl, alt)),
+        (Keycode::W, shift, ctrl, alt) => Some(("w", false, shift, ctrl, alt)),
+        (Keycode::X, shift, ctrl, alt) => Some(("x", false, shift, ctrl, alt)),
+        (Keycode::Y, shift, ctrl, alt) => Some(("y", false, shift, ctrl, alt)),
+        (Keycode::Z, shift, ctrl, alt) => Some(("z", false, shift, ctrl, alt)),
+        (Keycode::Delete, shift, ctrl, alt) => Some(("Delete", true, shift, ctrl, alt)),
+        (Keycode::F1, shift, ctrl, alt) => Some(("F1", true, shift, ctrl, alt)),
+        (Keycode::F2, shift, ctrl, alt) => Some(("F2", true, shift, ctrl, alt)),
+        (Keycode::F3, shift, ctrl, alt) => Some(("F3", true, shift, ctrl, alt)),
+        (Keycode::F4, shift, ctrl, alt) => Some(("F4", true, shift, ctrl, alt)),
+        (Keycode::F5, shift, ctrl, alt) => Some(("F5", true, shift, ctrl, alt)),
+        (Keycode::F6, shift, ctrl, alt) => Some(("F6", true, shift, ctrl, alt)),
+        (Keycode::F7, shift, ctrl, alt) => Some(("F7", true, shift, ctrl, alt)),
+        (Keycode::F8, shift, ctrl, alt) => Some(("F8", true, shift, ctrl, alt)),
+        (Keycode::F9, shift, ctrl, alt) => Some(("F9", true, shift, ctrl, alt)),
+        (Keycode::F10, shift, ctrl, alt) => Some(("F10", true, shift, ctrl, alt)),
+        (Keycode::F11, shift, ctrl, alt) => Some(("F11", true, shift, ctrl, alt)),
+        (Keycode::F12, shift, ctrl, alt) => Some(("F12", true, shift, ctrl, alt)),
+        (Keycode::Snapshot, _, _, _) => unsupported_key(Keycode::Snapshot),
+        (Keycode::Pause, _, _, _) => unsupported_key(Keycode::Pause),
+        (Keycode::Insert, shift, ctrl, alt) => Some(("Insert", true, shift, ctrl, alt)),
+        (Keycode::Home, shift, ctrl, alt) => Some(("Home", true, shift, ctrl, alt)),
+        (Keycode::PageUp, shift, ctrl, alt) => Some(("PageUp", true, shift, ctrl, alt)),
+        (Keycode::End, shift, ctrl, alt) => Some(("End", true, shift, ctrl, alt)),
+        (Keycode::PageDown, shift, ctrl, alt) => Some(("PageDown", true, shift, ctrl, alt)),
+        (Keycode::Right, shift, ctrl, alt) => Some(("Right", true, shift, ctrl, alt)),
+        (Keycode::Left, shift, ctrl, alt) => Some(("Left", true, shift, ctrl, alt)),
+        (Keycode::Down, shift, ctrl, alt) => Some(("Down", true, shift, ctrl, alt)),
+        (Keycode::Up, shift, ctrl, alt) => Some(("Up", true, shift, ctrl, alt)),
+        (Keycode::Numpad0, shift, ctrl, alt) => Some(("0", false, shift, ctrl, alt)),
+        (Keycode::Numpad1, shift, ctrl, alt) => Some(("1", false, shift, ctrl, alt)),
+        (Keycode::Numpad2, shift, ctrl, alt) => Some(("2", false, shift, ctrl, alt)),
+        (Keycode::Numpad3, shift, ctrl, alt) => Some(("3", false, shift, ctrl, alt)),
+        (Keycode::Numpad4, shift, ctrl, alt) => Some(("4", false, shift, ctrl, alt)),
+        (Keycode::Numpad5, shift, ctrl, alt) => Some(("5", false, shift, ctrl, alt)),
+        (Keycode::Numpad6, shift, ctrl, alt) => Some(("6", false, shift, ctrl, alt)),
+        (Keycode::Numpad7, shift, ctrl, alt) => Some(("7", false, shift, ctrl, alt)),
+        (Keycode::Numpad8, shift, ctrl, alt) => Some(("8", false, shift, ctrl, alt)),
+        (Keycode::Numpad9, shift, ctrl, alt) => Some(("9", false, shift, ctrl, alt)),
+        (Keycode::Apps, _, _, _) => unsupported_key(Keycode::Apps),
+        (Keycode::Power, _, _, _) => unsupported_key(Keycode::Power),
+        (Keycode::F13, shift, ctrl, alt) => Some(("F13", true, shift, ctrl, alt)),
+        (Keycode::F14, shift, ctrl, alt) => Some(("F14", true, shift, ctrl, alt)),
+        (Keycode::F15, shift, ctrl, alt) => Some(("F15", true, shift, ctrl, alt)),
+        (Keycode::F16, shift, ctrl, alt) => Some(("F16", true, shift, ctrl, alt)),
+        (Keycode::F17, shift, ctrl, alt) => Some(("F17", true, shift, ctrl, alt)),
+        (Keycode::F18, shift, ctrl, alt) => Some(("F18", true, shift, ctrl, alt)),
+        (Keycode::F19, shift, ctrl, alt) => Some(("F19", true, shift, ctrl, alt)),
+        (Keycode::F20, shift, ctrl, alt) => Some(("F20", true, shift, ctrl, alt)),
+        (Keycode::F21, shift, ctrl, alt) => Some(("F21", true, shift, ctrl, alt)),
+        (Keycode::F22, shift, ctrl, alt) => Some(("F22", true, shift, ctrl, alt)),
+        (Keycode::F23, shift, ctrl, alt) => Some(("F23", true, shift, ctrl, alt)),
+        (Keycode::F24, shift, ctrl, alt) => Some(("F24", true, shift, ctrl, alt)),
+        (Keycode::Cut, _, _, _) => unsupported_key(Keycode::Cut),
+        (Keycode::Copy, _, _, _) => unsupported_key(Keycode::Copy),
+        (Keycode::Paste, _, _, _) => unsupported_key(Keycode::Paste),
+        (Keycode::Mute, _, _, _) => unsupported_key(Keycode::Mute),
+        (Keycode::VolumeUp, _, _, _) => unsupported_key(Keycode::VolumeUp),
+        (Keycode::VolumeDown, _, _, _) => unsupported_key(Keycode::VolumeDown),
+        (Keycode::Sysrq, _, _, _) => unsupported_key(Keycode::Sysrq),
+        (Keycode::LControl, _, _, _) => None,
+        (Keycode::LShift, _, _, _) => None,
+        (Keycode::LAlt, _, _, _) => None,
+        (Keycode::RControl, _, _, _) => None,
+        (Keycode::RShift, _, _, _) => None,
+        (Keycode::RAlt, _, _, _) => None,
+        (Keycode::MediaStop, _, _, _) => unsupported_key(Keycode::MediaStop),
+        (Keycode::MediaSelect, _, _, _) => unsupported_key(Keycode::MediaSelect),
+        (Keycode::Calculator, _, _, _) => unsupported_key(Keycode::Calculator),
+        (Keycode::MyComputer, _, _, _) => unsupported_key(Keycode::MyComputer),
+        (Keycode::WebSearch, _, _, _) => unsupported_key(Keycode::WebSearch),
+        (Keycode::WebHome, _, _, _) => unsupported_key(Keycode::WebHome),
+        (Keycode::WebBack, _, _, _) => unsupported_key(Keycode::WebBack),
+        (Keycode::WebForward, _, _, _) => unsupported_key(Keycode::WebForward),
+        (Keycode::NavigateBackward, _, _, _) => unsupported_key(Keycode::NavigateBackward),
+        (Keycode::NavigateForward, _, _, _) => unsupported_key(Keycode::NavigateForward),
+        (Keycode::WebStop, _, _, _) => unsupported_key(Keycode::WebStop),
+        (Keycode::WebRefresh, _, _, _) => unsupported_key(Keycode::WebRefresh),
+        (Keycode::Sleep, _, _, _) => unsupported_key(Keycode::Sleep),
+        _ => None,
+    }
+}
